updated files

+++ ft_strlcat.c
@@ -72,12 +72,13 @@ unsigned int    ft_strlcat(char *dest, char *src, unsigned int size)
 	return (dest_len + ft_strlen(src)); 
 } 
 */
-#include <stddef.h>
+// #include <stddef.h>
+#include "libft.h"
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
